pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ikki0006", email = "ikki0006@gmail.com" }
]
dependencies = [
    "boto3-stubs>=1.35.74",
    "boto3>=1.35.74",
    "fastapi>=0.115.4",
    "pydantic-settings>=2.6.1",
    "uvicorn[standard]>=0.32.0",
    "python-ulid>=3.0.0",
    "anthropic>=0.40.0",
    "jinja2>=3.1.4",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.2",
    "mypy>=1.13.0",
    "pydantic[mypy]>=2.9.2",
    "pytest>=8.3.4",
    "httpx>=0.28.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
python_version = "3.12"
show_error_context = true  # エラー時のメッセージを詳細表示
show_column_numbers = true  # エラー発生箇所の行数/列数を表示
disallow_untyped_defs = true  # 関数定義の引数/戻り値に型アノテーション必須
no_implicit_optional = true  # デフォルト引数に None を取る場合型アノテーションに Optional 必須
check_untyped_defs = true  # 型注釈がない関数やメソッドに対して型チェックを行う
warn_redundant_casts = true  # 冗長なキャストに警告
warn_return_any = true  # 戻り値の型が Any になっている場合に警告
disallow_any_explicit = true  # 明示的な Any 型の使用を禁止
disallow_untyped_calls = true  # 型アノテーションのない関数呼び出しを禁止
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "ulid.*",
    "httpx.*",
    "anthropic.*",
    "jinja2.*"
]
ignore_missing_imports = true

[tool.ruff]


# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
